test(side-menu): run the pure rect arithmetic outside the browser
`rectsAreSideBySide` and `rectIndexAtCursor` take plain rects and return
plain answers, but their tests sat in the `.browser.test` file and so paid
for a browser to run arithmetic. They move to a node test; the walks that
measure live layout (`getBoundingClientRect`, real flex) stay behind.

`containerParse.browser.test.ts` was considered too and stays as it is:
it exercises HTML parsing, so validating it against jsdom's parser rather
than a real engine's would defeat the test.
